refactor: 添付ファイル共通処理を src/shared/attachment.ts に抽出
Wiki・ドキュメントへの拡張に先立ち、課題実装から attachmentFileName /
encodeLinkDestination / escapeLinkText / rewriteInlineImages を共有化する。
buildIssueExpectedPaths の添付パス登録は、pruneがattachmentsディレクトリを
スキップする方式になり不要になったため削除。
