+++ src/modules/issue/domain/issue-markdown.ts
@@ -1,3 +1,4 @@
+import {escapeLinkText, rewriteInlineImages} from '../../../shared/attachment.js'
 import {wrapBody} from '../../../shared/markdown/body-marker.js'
 import {CustomField, Issue, IssueAttachment, IssueComment, IssueCommentChange} from './issue.js'
 
@@ -33,38 +34,6 @@ function formatChangeValue(value: null | string): string {
 function formatChange(change: IssueCommentChange): string {
   const label = CHANGE_FIELD_LABELS[change.field] ?? change.field
   return `- ${label}: ${formatChangeValue(change.originalValue)} → ${formatChangeValue(change.newValue)}`
-}
-
-// Backlogの添付画像インライン記法（Markdown拡張の ![alt][ファイル名] とBacklog記法の #image(ファイル名)）を
-// ダウンロード済みファイルへのローカルリンクに変換する。未ダウンロードの参照は壊さずそのまま残す
-export function rewriteInlineImages(
-  text: string,
-  attachments: IssueAttachment[] | undefined,
-  localLinks?: Map<number, string>,
-): string {
-  if (!text || !attachments || attachments.length === 0 || !localLinks || localLinks.size === 0) {
-    return text
-  }
-
-  // 同名添付が複数ある場合は記法から特定できないため先勝ちで解決する。
-  // 記法内のファイル名はNFD（macOSからのD&D等）で入ることがあるため、照合はNFC正規化で行う
-  const linkByName = new Map<string, string>()
-  for (const attachment of attachments) {
-    const link = localLinks.get(attachment.id)
-    const name = attachment.name.normalize('NFC')
-    if (link && !linkByName.has(name)) {
-      linkByName.set(name, link)
-    }
-  }
-
-  const toLocalImage = (match: string, name: string) => {
-    const link = linkByName.get(name.normalize('NFC'))
-    return link ? `![${escapeLinkText(name)}](${link})` : match
-  }
-
-  return text
-    .replaceAll(/!\[[^\]]*\]\[([^\]]+)\]/g, toLocalImage)
-    .replaceAll(/#image\(([^)]+)\)/g, toLocalImage)
 }
 
 function buildCommentBody(comment: IssueComment, rewriteBody: (text: string) => string): string {
@@ -113,11 +82,6 @@ export function createCustomFieldsSection(customFields?: CustomField[]): string 
   }
 
   return customFieldsSection
-}
-
-// ファイル名中の角括弧はリンク構文を壊すためエスケープする
-function escapeLinkText(name: string): string {
-  return name.replaceAll('[', String.raw`\[`).replaceAll(']', String.raw`\]`)
 }
 
 // ダウンロード済みの添付はローカルへの相対リンク付き、未ダウンロードはメタデータのみを出力する
