+++ src/modules/issue/domain/issue-path.ts
@@ -1,7 +1,8 @@
 import path from 'node:path'
 
+import {attachmentFileName, encodeLinkDestination} from '../../../shared/attachment.js'
 import {backlogOrigin} from '../../../shared/backlog-url.js'
-import {sanitizeAttachmentFileName, sanitizeFileName} from '../../../shared/file-name.js'
+import {sanitizeFileName} from '../../../shared/file-name.js'
 import {ExpectedPaths} from '../../prune/domain/expected-paths.js'
 import {Issue, IssueAttachment} from './issue.js'
 
@@ -24,11 +25,6 @@ export function issueRelativePath(
   )
 }
 
-// 添付IDを前置して同一課題内の同名添付の衝突を防ぎ、存在チェックだけでDL済み判定できるようにする
-export function attachmentFileName(attachment: Pick<IssueAttachment, 'id' | 'name'>): string {
-  return `${attachment.id}_${sanitizeAttachmentFileName(attachment.name)}`
-}
-
 // issueKeyFolderありなら課題フォルダ直下のattachments/、なしなら年フォルダのattachments/{課題キー}/
 function attachmentDirSegments(issueKey: string, useIssueKeyFolder: boolean): string[] {
   return useIssueKeyFolder ? ['attachments'] : ['attachments', issueKey]
@@ -46,14 +42,13 @@ export function attachmentRelativePath(
   )
 }
 
-// Markdownファイルから添付への相対リンク。丸括弧はインラインリンクを壊すためエンコードする
 export function attachmentMarkdownLink(
   issue: {issueKey: string},
   attachment: Pick<IssueAttachment, 'id' | 'name'>,
   options: {issueKeyFolder?: boolean},
 ): string {
   const segments = attachmentDirSegments(issue.issueKey, options.issueKeyFolder ?? false)
-  return ['.', ...segments, attachmentFileName(attachment)].join('/').replaceAll('(', '%28').replaceAll(')', '%29')
+  return encodeLinkDestination(['.', ...segments, attachmentFileName(attachment)].join('/'))
 }
 
 export function issueUrl(domain: string, issueKey: string): string {
@@ -68,23 +63,13 @@ export function buildIssueExpectedPaths(
   const expectedFiles = new Set<string>()
   const expectedDirs = new Set<string>()
 
-  const addDirs = (dir: string) => {
-    let current = dir
-    while (current && current !== '.') {
-      expectedDirs.add(current.normalize('NFC'))
-      current = path.dirname(current)
-    }
-  }
-
   for (const issue of issues) {
     expectedFiles.add(issueRelativePath(issue, options).normalize('NFC'))
-    addDirs(issueRelativeDir(issue, options.issueKeyFolder ?? false))
 
-    // .md拡張子の添付ファイルがpruneで誤削除されないよう、添付パスも期待集合に含める
-    for (const attachment of issue.attachments ?? []) {
-      const attachmentPath = attachmentRelativePath(issue, attachment, options)
-      expectedFiles.add(attachmentPath.normalize('NFC'))
-      addDirs(path.dirname(attachmentPath))
+    let dir = issueRelativeDir(issue, options.issueKeyFolder ?? false)
+    while (dir && dir !== '.') {
+      expectedDirs.add(dir.normalize('NFC'))
+      dir = path.dirname(dir)
     }
   }
